PCRE: moved structure definition to header.
Logging the patterns tried in the router requires being able to access
the fields of this structure.
